test: split the sample specs and share the document order helper
